We back with a UI update.
Got images (and file name) to show up when you hit the process button after selecting a file. Not the way I want them to show up yet, was hoping to have them show on the previous screen when the browse dialog box exits.

Created image objects, each image has the actual file, and the location on your computer. They are stored on an global arrayList as you select them (No way to currently navigate through your list).
So when we process a new image, we can store it, or possibly access it with other parts of our code. I needed to do this to access them after bowsing.

Added sick PicPro image to main menu. (needs work)